+++ src/test/java/Encriptador/controller/EncriptadorControllerTest.java
@@ -7,13 +7,14 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class EncriptadorControllerTest {
     EncriptadorController encriptador = new EncriptadorController();
+
     @Test
     @DisplayName("Generar Matriz a partir del mensaje hola")
     void generarMatrizPrueba1() {
         String mensaje = "hola";
         int columnas = encriptador.obtenerColumnas(mensaje);
-        int [][] matrizGenerada = encriptador.generarMatriz(mensaje, columnas);
-        int [][] matrizResultado = {{8,1},
+        double [][] matrizGenerada = encriptador.generarMatriz(mensaje, columnas);
+        double [][] matrizResultado = {{8,1},
                                     {15,0},
                                     {12,0}};
 
@@ -30,8 +31,8 @@ class EncriptadorControllerTest {
     void generarMatrizPrueba2() {
         String mensaje = "hola que tal amix";
         int columnas = encriptador.obtenerColumnas(mensaje);
-        int [][] matrizGenerada = encriptador.generarMatriz(mensaje, columnas);
-        int [][] matrizResultado = {{8,1,21,20,0,9},
+        double [][] matrizGenerada = encriptador.generarMatriz(mensaje, columnas);
+        double [][] matrizResultado = {{8,1,21,20,0,9},
                                     {15,0,5,1,1,24},
                                     {12,17,0,12,13,0}};
         for(int fila=0;fila<3;fila++){
@@ -45,7 +46,7 @@ class EncriptadorControllerTest {
     @Test
     @DisplayName("Generar Llave Aleatoria")
     void generarLlave(){
-        int [][] llaveAleatoria = encriptador.generarLlave();
+        double [][] llaveAleatoria = encriptador.generarLlave();
         for(int fila=0;fila<3;fila++){
             for(int columna=0;columna<3;columna++){
                 System.out.print(llaveAleatoria[fila][columna]+" ");
@@ -60,7 +61,6 @@ class EncriptadorControllerTest {
     void encriptarMensaje(){
         EncriptadorController encriptar = new EncriptadorController();
         encriptar.encriptarMensaje();
-
     }
